# feat: pass path to censor function