--- a/lib/modifiers.js
+++ b/lib/modifiers.js
@@ -20,12 +20,15 @@ function groupRedact (o, path, censor, isCensorFct) {
   const target = get(o, path)
   if (target == null) return { keys: null, values: null, target: null, flat: true }
   const keys = Object.keys(target)
-  const length = keys.length
-  const values = new Array(length)
-  for (var i = 0; i < length; i++) {
-    const k = keys[i]
-    values[i] = target[k]
-    target[k] = isCensorFct ? censor(target[k]) : censor
+  const keysLength = keys.length
+  const pathLength = path.length
+  const pathWithKey = path.concat('')
+  const values = new Array(keysLength)
+  for (var i = 0; i < keysLength; i++) {
+    const key = keys[i]
+    pathWithKey[pathLength] = key
+    values[i] = target[key]
+    target[key] = isCensorFct ? censor(target[key], pathWithKey) : censor
   }
   return { keys, values, target, flat: true }
 }
@@ -42,10 +45,13 @@ function nestedRedact (store, o, path, ns, censor, isCensorFct) {
   const target = get(o, path)
   if (target == null) return
   const keys = Object.keys(target)
-  const length = keys.length
-  for (var i = 0; i < length; i++) {
+  const keysLength = keys.length
+  const pathLength = path.length
+  const pathWithKey = path.concat('')
+  for (var i = 0; i < keysLength; i++) {
     const key = keys[i]
-    const { value, parent, exists } = specialSet(target, key, ns, censor, isCensorFct)
+    pathWithKey[pathLength] = key
+    const { value, parent, exists } = specialSet(target, key, pathWithKey, ns, censor, isCensorFct)
 
     if (exists === true && parent !== null) {
       store.push({ key: ns[ns.length - 1], target: parent, value })
@@ -58,10 +64,10 @@ function has (obj, prop) {
   return Object.prototype.hasOwnProperty.call(obj, prop)
 }
 
-function specialSet (o, k, p, v, f) {
+function specialSet (o, k, pathWithKey, afterPath, censor, isCensorFct) {
+  const afterPathLen = afterPath.length
+  const lastPathIndex = afterPathLen - 1
   var i = -1
-  var l = p.length
-  var li = l - 1
   var n
   var nv
   var ov
@@ -69,22 +75,26 @@ function specialSet (o, k, p, v, f) {
   var exists = true
   ov = n = o[k]
   if (typeof n !== 'object') return { value: null, parent: null, exists }
-  while (n != null && ++i < l) {
-    k = p[i]
+  while (n != null && ++i < afterPathLen) {
+    k = afterPath[i]
     oov = ov
     if (!(k in n)) {
       exists = false
       break
     }
     ov = n[k]
-    nv = f ? v(ov) : v
-    nv = (i !== li) ? ov : nv
-    n[k] = (has(n, k) && nv === ov) || (nv === undefined && v !== undefined) ? n[k] : nv
+    nv = (i !== lastPathIndex)
+      ? ov
+      : (isCensorFct
+        ? censor(ov, pathWithKey.concat(afterPath))
+        : censor)
+    n[k] = (has(n, k) && nv === ov) || (nv === undefined && censor !== undefined) ? n[k] : nv
     n = n[k]
     if (typeof n !== 'object') break
   }
   return { value: ov, parent: oov, exists }
 }
+
 function get (o, p) {
   var i = -1
   var l = p.length