--- a/lib/modifiers.js
+++ b/lib/modifiers.js
@@ -22,13 +22,19 @@ function groupRedact (o, path, censor, isCensorFct) {
   const keys = Object.keys(target)
   const keysLength = keys.length
   const pathLength = path.length
-  const pathWithKey = path.concat('')
+  const pathWithKey = isCensorFct ? [...path, ''] : undefined
   const values = new Array(keysLength)
+
   for (var i = 0; i < keysLength; i++) {
     const key = keys[i]
-    pathWithKey[pathLength] = key
     values[i] = target[key]
-    target[key] = isCensorFct ? censor(target[key], pathWithKey) : censor
+
+    if (isCensorFct) {
+      pathWithKey[pathLength] = key
+      target[key] = censor(target[key], pathWithKey)
+    } else {
+      target[key] = censor
+    }
   }
   return { keys, values, target, flat: true }
 }
@@ -46,12 +52,10 @@ function nestedRedact (store, o, path, ns, censor, isCensorFct) {
   if (target == null) return
   const keys = Object.keys(target)
   const keysLength = keys.length
-  const pathLength = path.length
-  const pathWithKey = path.concat('')
   for (var i = 0; i < keysLength; i++) {
     const key = keys[i]
-    pathWithKey[pathLength] = key
-    const { value, parent, exists } = specialSet(target, key, pathWithKey, ns, censor, isCensorFct)
+    const { value, parent, exists } =
+      specialSet(target, key, path, ns, censor, isCensorFct)
 
     if (exists === true && parent !== null) {
       store.push({ key: ns[ns.length - 1], target: parent, value })
@@ -64,9 +68,10 @@ function has (obj, prop) {
   return Object.prototype.hasOwnProperty.call(obj, prop)
 }
 
-function specialSet (o, k, pathWithKey, afterPath, censor, isCensorFct) {
+function specialSet (o, k, path, afterPath, censor, isCensorFct) {
   const afterPathLen = afterPath.length
   const lastPathIndex = afterPathLen - 1
+  const originalKey = k
   var i = -1
   var n
   var nv
@@ -86,7 +91,7 @@ function specialSet (o, k, pathWithKey, afterPath, censor, isCensorFct) {
     nv = (i !== lastPathIndex)
       ? ov
       : (isCensorFct
-        ? censor(ov, pathWithKey.concat(afterPath))
+        ? censor(ov, [...path, originalKey, ...afterPath])
         : censor)
     n[k] = (has(n, k) && nv === ov) || (nv === undefined && censor !== undefined) ? n[k] : nv
     n = n[k]