--- a/lib/modifiers.js
+++ b/lib/modifiers.js
@@ -22,7 +22,7 @@ function groupRedact (o, path, censor, isCensorFct) {
   const keys = Object.keys(target)
   const keysLength = keys.length
   const pathLength = path.length
-  const pathWithKey = isCensorFct ? [...path, ''] : undefined
+  const pathWithKey = isCensorFct ? [...path] : undefined
   const values = new Array(keysLength)
 
   for (var i = 0; i < keysLength; i++) {