--- a/lib/modifiers.js
+++ b/lib/modifiers.js
@@ -16,22 +16,24 @@ function groupRestore ({ keys, values, target }) {
   }
 }
 
-function groupRedact (o, path, censor, isCensorFct) {
+function groupRedact (o, path, censor, isCensorFct, censorFctTakesPath) {
   const target = get(o, path)
   if (target == null) return { keys: null, values: null, target: null, flat: true }
   const keys = Object.keys(target)
   const keysLength = keys.length
   const pathLength = path.length
-  const pathWithKey = isCensorFct ? [...path] : undefined
+  const pathWithKey = censorFctTakesPath ? [...path] : undefined
   const values = new Array(keysLength)
 
   for (var i = 0; i < keysLength; i++) {
     const key = keys[i]
     values[i] = target[key]
 
-    if (isCensorFct) {
+    if (censorFctTakesPath) {
       pathWithKey[pathLength] = key
       target[key] = censor(target[key], pathWithKey)
+    } else if (isCensorFct) {
+      target[key] = censor(target[key])
     } else {
       target[key] = censor
     }
@@ -47,7 +49,7 @@ function nestedRestore (arr) {
   }
 }
 
-function nestedRedact (store, o, path, ns, censor, isCensorFct) {
+function nestedRedact (store, o, path, ns, censor, isCensorFct, censorFctTakesPath) {
   const target = get(o, path)
   if (target == null) return
   const keys = Object.keys(target)
@@ -55,7 +57,7 @@ function nestedRedact (store, o, path, ns, censor, isCensorFct) {
   for (var i = 0; i < keysLength; i++) {
     const key = keys[i]
     const { value, parent, exists } =
-      specialSet(target, key, path, ns, censor, isCensorFct)
+      specialSet(target, key, path, ns, censor, isCensorFct, censorFctTakesPath)
 
     if (exists === true && parent !== null) {
       store.push({ key: ns[ns.length - 1], target: parent, value })
@@ -68,7 +70,7 @@ function has (obj, prop) {
   return Object.prototype.hasOwnProperty.call(obj, prop)
 }
 
-function specialSet (o, k, path, afterPath, censor, isCensorFct) {
+function specialSet (o, k, path, afterPath, censor, isCensorFct, censorFctTakesPath) {
   const afterPathLen = afterPath.length
   const lastPathIndex = afterPathLen - 1
   const originalKey = k
@@ -91,7 +93,7 @@ function specialSet (o, k, path, afterPath, censor, isCensorFct) {
     nv = (i !== lastPathIndex)
       ? ov
       : (isCensorFct
-        ? censor(ov, [...path, originalKey, ...afterPath])
+        ? (censorFctTakesPath ? censor(ov, [...path, originalKey, ...afterPath]) : censor(ov))
         : censor)
     n[k] = (has(n, k) && nv === ov) || (nv === undefined && censor !== undefined) ? n[k] : nv
     n = n[k]