--- a/lib/redactor.js
+++ b/lib/redactor.js
@@ -26,7 +26,7 @@ function redactor ({ secret, serialize, wcLen, strict, isCensorFct }, state) {
 
 function redactTmpl (secret, isCensorFct) {
   return Object.keys(secret).map((path) => {
-    const { escPath, leadingBracket } = secret[path]
+    const { escPath, leadingBracket, path: arrPath } = secret[path]
     const skip = leadingBracket ? 1 : 0
     const delim = leadingBracket ? '' : '.'
     const hops = []
@@ -56,7 +56,7 @@ function redactTmpl (secret, isCensorFct) {
           secret[${escPath}].precensored = true
         } else {
           secret[${escPath}].val = val
-          o${delim}${path} = ${isCensorFct ? 'censor(val)' : 'censor'}
+          o${delim}${path} = ${isCensorFct ? `censor(val, ${JSON.stringify(arrPath)})` : 'censor'}
           ${circularDetection}
         }
       }