--- a/lib/redactor.js
+++ b/lib/redactor.js
@@ -4,7 +4,7 @@ const rx = require('./rx')
 
 module.exports = redactor
 
-function redactor ({ secret, serialize, wcLen, strict, isCensorFct }, state) {
+function redactor ({ secret, serialize, wcLen, strict, isCensorFct, censorFctTakesPath }, state) {
   /* eslint-disable-next-line */
   const redact = Function('o', `
     if (typeof o !== 'object' || o == null) {
@@ -13,7 +13,7 @@ function redactor ({ secret, serialize, wcLen, strict, isCensorFct }, state) {
     const { censor, secret } = this
     ${redactTmpl(secret, isCensorFct)}
     this.compileRestore()
-    ${dynamicRedactTmpl(wcLen > 0, isCensorFct)}
+    ${dynamicRedactTmpl(wcLen > 0, isCensorFct, censorFctTakesPath)}
     ${resultTmpl(serialize)}
   `).bind(state)
 
@@ -64,7 +64,7 @@ function redactTmpl (secret, isCensorFct) {
   }).join('\n')
 }
 
-function dynamicRedactTmpl (hasWildcards, isCensorFct) {
+function dynamicRedactTmpl (hasWildcards, isCensorFct, censorFctTakesPath) {
   return hasWildcards === true ? `
     {
       const { wildcards, wcLen, groupRedact, nestedRedact } = this
@@ -72,8 +72,8 @@ function dynamicRedactTmpl (hasWildcards, isCensorFct) {
         const { before, beforeStr, after, nested } = wildcards[i]
         if (nested === true) {
           secret[beforeStr] = secret[beforeStr] || []
-          nestedRedact(secret[beforeStr], o, before, after, censor, ${isCensorFct})
-        } else secret[beforeStr] = groupRedact(o, before, censor, ${isCensorFct})
+          nestedRedact(secret[beforeStr], o, before, after, censor, ${isCensorFct}, ${censorFctTakesPath})
+        } else secret[beforeStr] = groupRedact(o, before, censor, ${isCensorFct}, ${censorFctTakesPath})
       }
     }
   ` : ''