--- a/lib/redactor.js
+++ b/lib/redactor.js
@@ -11,7 +11,7 @@ function redactor ({ secret, serialize, wcLen, strict, isCensorFct, censorFctTak
       ${strictImpl(strict, serialize)}
     }
     const { censor, secret } = this
-    ${redactTmpl(secret, isCensorFct)}
+    ${redactTmpl(secret, isCensorFct, censorFctTakesPath)}
     this.compileRestore()
     ${dynamicRedactTmpl(wcLen > 0, isCensorFct, censorFctTakesPath)}
     ${resultTmpl(serialize)}
@@ -24,7 +24,7 @@ function redactor ({ secret, serialize, wcLen, strict, isCensorFct, censorFctTak
   return redact
 }
 
-function redactTmpl (secret, isCensorFct) {
+function redactTmpl (secret, isCensorFct, censorFctTakesPath) {
   return Object.keys(secret).map((path) => {
     const { escPath, leadingBracket, path: arrPath } = secret[path]
     const skip = leadingBracket ? 1 : 0
@@ -49,6 +49,11 @@ function redactTmpl (secret, isCensorFct) {
         `).join('\n')}
       }
     `
+
+    const censorArgs = censorFctTakesPath
+      ? `val, ${JSON.stringify(arrPath)}`
+      : `val`
+
     return `
       if (${existence}) {
         const val = o${delim}${path}
@@ -56,7 +61,7 @@ function redactTmpl (secret, isCensorFct) {
           secret[${escPath}].precensored = true
         } else {
           secret[${escPath}].val = val
-          o${delim}${path} = ${isCensorFct ? `censor(val, ${JSON.stringify(arrPath)})` : 'censor'}
+          o${delim}${path} = ${isCensorFct ? `censor(${censorArgs})` : 'censor'}
           ${circularDetection}
         }
       }