--- a/lib/state.js
+++ b/lib/state.js
@@ -6,7 +6,6 @@ function state (o) {
   const {
     secret,
     censor,
-    isCensorFct,
     compileRestore,
     serialize,
     groupRedact,
@@ -14,8 +13,7 @@ function state (o) {
     wildcards,
     wcLen
   } = o
-  const builder = [{ secret, censor, isCensorFct, compileRestore }]
-  builder.push({ secret })
+  const builder = [{ secret, censor, compileRestore }]
   if (serialize !== false) builder.push({ serialize })
   if (wcLen > 0) builder.push({ groupRedact, nestedRedact, wildcards, wcLen })
   return Object.assign(...builder)