--- a/package.json
+++ b/package.json
@@ -1,6 +1,6 @@
 {
   "name": "fast-redact",
-  "version": "1.4.3",
+  "version": "1.4.4",
   "description": "very fast object redaction",
   "main": "index.js",
   "scripts": {