--- a/test/index.js
+++ b/test/index.js
@@ -4,7 +4,7 @@ const { test } = require('tap')
 const fastRedact = require('..')
 
 const censor = '[REDACTED]'
-const censorFct = value => !value ? value : 'xxx' + value.substr(-2)
+const censorFct = (value, path) => !value ? value : 'xxx' + value.substr(-2) + path.join('.')
 
 test('returns no-op when passed no paths [serialize: false]', ({ end, doesNotThrow }) => {
   const redact = fastRedact({ paths: [], serialize: false })
@@ -230,20 +230,20 @@ test('redact.restore function places original values back in place', ({ end, is 
 
 test('masks according to supplied censor function', ({ end, is }) => {
   const redact = fastRedact({ paths: ['a'], censor: censorFct, serialize: false })
-  is(redact({ a: '0123456' }).a, 'xxx56')
+  is(redact({ a: '0123456' }).a, 'xxx56a')
   end()
 })
 
 test('masks according to supplied censor function with wildcards', ({ end, is }) => {
   const redact = fastRedact({ paths: '*', censor: censorFct, serialize: false })
-  is(redact({ a: '0123456' }).a, 'xxx56')
+  is(redact({ a: '0123456' }).a, 'xxx56a')
   end()
 })
 
 test('masks according to supplied censor function with nested wildcards', ({ end, is }) => {
   const redact = fastRedact({ paths: ['*.b'], censor: censorFct, serialize: false })
-  is(redact({ a: { b: '0123456' } }).a.b, 'xxx56')
-  is(redact({ c: { b: '0123456', d: 'pristine' } }).c.b, 'xxx56')
+  is(redact({ a: { b: '0123456' } }).a.b, 'xxx56a.b')
+  is(redact({ c: { b: '0123456', d: 'pristine' } }).c.b, 'xxx56c.b')
   is(redact({ c: { b: '0123456', d: 'pristine' } }).c.d, 'pristine')
   end()
 })
@@ -252,7 +252,7 @@ test('redact.restore function places original values back in place with censor f
   const redact = fastRedact({ paths: ['a'], censor: censorFct, serialize: false })
   const o = { a: 'qwerty' }
   redact(o)
-  is(o.a, 'xxxty')
+  is(o.a, 'xxxtya')
   redact.restore(o)
   is(o.a, 'qwerty')
   end()