This PR calls the user's censor function with the path being censored as the second argument. I'm trying to support two different use cases here:

1. Sometimes different keys need to be redacted in different ways. For example, on an object representing a bank account, you might censor the account number by `xxxx`-ing out most of the numbers (but still leave a few for debugging), while censoring the account owner's name by removing it completely:

    ```js
    censor: (v, path) => path[path.length - 1] === 'accountNumber' 
      ? `xxxxxx${v.substr(-4)}` 
      : '[Redacted]'
    ```

2. Because fast-redact serializes the redacted object rather than returning a deep clone (for well-explained reasons), it's difficult to compose redaction logic. For example, suppose I have a few different business objects, like bank accounts, clients, etc., and that I want to define a separate redaction function for each object (in the same place where I define the object's schema etc). For example, I might define `redactAccount` as:
    ```js
    const censorAccount; // same censor function shown above
    const accountSecretPaths = ["accountNumber", "otherSecretPath"];
    const redactAccount = fastRedact({
      paths: accountSecretPaths, 
      censor: censorAccount
    });
    ```

    Now, suppose I need to log a redacted version of an object like: 

    ```js
    { clients: [...], accounts: [...] }
    ```

     If my `redactAccount` et al functions returned censored objects, rather than strings, this would be as simple as: 

    ```js
    log({ clients: [...].map(redactClient), accounts: [...].map(redactAccount) })
    ```

    So, my thought is that, by passing the path to `censor` functions, it becomes possible to support cases like this. I.e., I can now define a `composedRedact` function like so:

    ```js
    const opts = {
      paths: [
        ...accountSecretPaths.map(it => `accounts.*.${it}`),
        ...clientSecretPaths.map(it => `clients.*.${it}`)
      ],
      censor: (v, path) => path[0] === 'clients' 
        ? censorClient(v, path.slice(2))
        : censorAccount(v, path.slice(2));
    };

    const composedRedact = fastRedact(opts);
    ```

     Fwiw, I explored approaches to composition that leveraged `serialize: false` and `restore()`, but ran up against the wall that the state stored internally for each redaction function returned by `fastRedact` only covers the last object redacted, so it wasn't possible to redact a bunch of objects without serialization, collect and serialize the results, and then restore all of them.

    Given that limitation, I couldn't think of an approach for supporting composition that, at least absent a major refactor, wouldn't, under the hood, do more or less what the above `opts.censor` and `opts.paths` are doing explicitly. (Although I certainly can imagine some more convenient APIs that could be added to abstract that basic logic.)