@ethanresnick 2019-03-18
Also, I'm certainly no V8 performance expert, but I tried to be relatively performance minded in my implementation here. Still, if its performance is an issue, I imagine that a relatively simple way to take down the overhead further would be to test the `.length` of `censor` function at the beginning and not bother with constructing the path array if the arity is only 1.