@ethanresnick 2019-03-18
@mcollina Thanks for the review. I'll try to work on this more tonight.