@ethanresnick 2019-03-18
As far as the benchmarks go, here's what I got:

### Before

```
benchNoirV2*500: 85.971ms
benchFastRedact*500: 41.433ms
benchFastRedactRestore*500: 14.223ms
benchNoirV2Wild*500: 67.439ms
benchFastRedactWild*500: 29.585ms
benchFastRedactWildRestore*500: 37.385ms
benchFastRedactIntermediateWild*500: 133.649ms
benchFastRedactIntermediateWildRestore*500: 93.075ms
benchJSONStringify*500: 236.068ms
benchNoirV2Serialize*500: 319.681ms
benchFastRedactSerialize*500: 240.579ms
benchNoirV2WildSerialize*500: 284.403ms
benchFastRedactWildSerialize*500: 276.943ms
benchFastRedactIntermediateWildSerialize*500: 328.257ms
benchFastRedactIntermediateWildMatchWildOutcomeSerialize*500: 477.666ms
benchFastRedactStaticMatchWildOutcomeSerialize*500: 262.410ms
benchNoirV2CensorFunction*500: 41.339ms
benchFastRedactCensorFunction*500: 67.306ms
benchNoirV2*500: 52.799ms
benchFastRedact*500: 2.785ms
benchFastRedactRestore*500: 15.687ms
benchNoirV2Wild*500: 93.922ms
benchFastRedactWild*500: 21.801ms
benchFastRedactWildRestore*500: 30.377ms
benchFastRedactIntermediateWild*500: 152.801ms
benchFastRedactIntermediateWildRestore*500: 94.812ms
benchJSONStringify*500: 234.192ms
benchNoirV2Serialize*500: 305.718ms
benchFastRedactSerialize*500: 235.601ms
benchNoirV2WildSerialize*500: 273.017ms
benchFastRedactWildSerialize*500: 269.712ms
benchFastRedactIntermediateWildSerialize*500: 308.397ms
benchFastRedactIntermediateWildMatchWildOutcomeSerialize*500: 470.652ms
benchFastRedactStaticMatchWildOutcomeSerialize*500: 265.717ms
benchNoirV2CensorFunction*500: 50.077ms
benchFastRedactCensorFunction*500: 45.616ms
```

### After
```
benchNoirV2*500: 85.883ms
benchFastRedact*500: 46.895ms
benchFastRedactRestore*500: 9.792ms
benchNoirV2Wild*500: 57.063ms
benchFastRedactWild*500: 29.695ms
benchFastRedactWildRestore*500: 36.148ms
benchFastRedactIntermediateWild*500: 122.937ms
benchFastRedactIntermediateWildRestore*500: 91.967ms
benchJSONStringify*500: 235.932ms
benchNoirV2Serialize*500: 322.515ms
benchFastRedactSerialize*500: 244.163ms
benchNoirV2WildSerialize*500: 292.856ms
benchFastRedactWildSerialize*500: 287.887ms
benchFastRedactIntermediateWildSerialize*500: 339.095ms
benchFastRedactIntermediateWildMatchWildOutcomeSerialize*500: 480.987ms
benchFastRedactStaticMatchWildOutcomeSerialize*500: 261.544ms
benchNoirV2CensorFunction*500: 76.100ms
benchFastRedactCensorFunction*500: 81.954ms
benchNoirV2*500: 87.294ms
benchFastRedact*500: 2.751ms
benchFastRedactRestore*500: 16.665ms
benchNoirV2Wild*500: 35.325ms
benchFastRedactWild*500: 20.978ms
benchFastRedactWildRestore*500: 28.181ms
benchFastRedactIntermediateWild*500: 152.879ms
benchFastRedactIntermediateWildRestore*500: 93.906ms
benchJSONStringify*500: 230.684ms
benchNoirV2Serialize*500: 307.351ms
benchFastRedactSerialize*500: 236.411ms
benchNoirV2WildSerialize*500: 272.602ms
benchFastRedactWildSerialize*500: 274.934ms
benchFastRedactIntermediateWildSerialize*500: 344.361ms
benchFastRedactIntermediateWildMatchWildOutcomeSerialize*500: 472.360ms
benchFastRedactStaticMatchWildOutcomeSerialize*500: 257.025ms
benchNoirV2CensorFunction*500: 70.302ms
benchFastRedactCensorFunction*500: 61.608ms
```