@ethanresnick 2019-10-09
> Could you also add a benchmark with the censor function and a benchmark the covers all code added to specialSet and post results?

@davidmarkclements I've pushed a commit that adds these benchmarks. There was already one benchmark using a censor function, but I've added a couple others that test the combination of a censor function (with and without the second `path` argument) and different redaction path wildcard scenarios. For the specialSet changes, the only new branch in the logic is the case of calling the censor function with a path, which is covered by the new benchmark. (All the rest of that function is covered by the existing benchmarks, I believe.)

Here are the results....

### Before 
These results are with none of the changes from this PR, but _with_ the rework to the benchmarks from #22. These are taken straight from the OP in #22.

```
benchNoirV2*500: 54.057ms
benchFastRedact*500: 15.949ms
benchFastRedactRestore*500: 13.749ms
benchNoirV2Wild*500: 34.787ms
benchFastRedactWild*500: 44.642ms
benchFastRedactWildRestore*500: 42.949ms
benchFastRedactIntermediateWild*500: 162.311ms
benchFastRedactIntermediateWildRestore*500: 106.921ms
benchJSONStringify*500: 306.766ms
benchNoirV2Serialize*500: 403.440ms
benchFastRedactSerialize*500: 321.294ms
benchNoirV2WildSerialize*500: 366.650ms
benchFastRedactWildSerialize*500: 342.946ms
benchFastRedactIntermediateWildSerialize*500: 437.797ms
benchFastRedactIntermediateWildMatchWildOutcomeSerialize*500: 573.556ms
benchFastRedactStaticMatchWildOutcomeSerialize*500: 336.521ms
benchNoirV2CensorFunction*500: 27.110ms
benchFastRedactCensorFunction*500: 47.454ms
benchNoirV2*500: 38.437ms
benchFastRedact*500: 10.272ms
benchFastRedactRestore*500: 9.693ms
benchNoirV2Wild*500: 18.504ms
benchFastRedactWild*500: 30.266ms
benchFastRedactWildRestore*500: 35.108ms
benchFastRedactIntermediateWild*500: 131.794ms
benchFastRedactIntermediateWildRestore*500: 110.691ms
benchJSONStringify*500: 299.861ms
benchNoirV2Serialize*500: 384.236ms
benchFastRedactSerialize*500: 314.049ms
benchNoirV2WildSerialize*500: 365.485ms
benchFastRedactWildSerialize*500: 344.158ms
benchFastRedactIntermediateWildSerialize*500: 426.421ms
benchFastRedactIntermediateWildMatchWildOutcomeSerialize*500: 537.079ms
benchFastRedactStaticMatchWildOutcomeSerialize*500: 340.104ms
benchNoirV2CensorFunction*500: 16.021ms
benchFastRedactCensorFunction*500: 31.100ms
```

### After
Same benchmarks, but with the code changes from this PR.

```
benchNoirV2*500: 57.173ms
benchFastRedact*500: 15.292ms
benchFastRedactRestore*500: 17.874ms
benchNoirV2Wild*500: 36.849ms
benchFastRedactWild*500: 42.018ms
benchFastRedactWildRestore*500: 43.830ms
benchFastRedactIntermediateWild*500: 170.819ms
benchFastRedactIntermediateWildRestore*500: 104.677ms
benchJSONStringify*500: 295.113ms
benchNoirV2Serialize*500: 400.466ms
benchFastRedactSerialize*500: 318.930ms
benchNoirV2WildSerialize*500: 358.291ms
benchFastRedactWildSerialize*500: 339.127ms
benchFastRedactIntermediateWildSerialize*500: 430.686ms
benchFastRedactIntermediateWildMatchWildOutcomeSerialize*500: 578.805ms
benchFastRedactStaticMatchWildOutcomeSerialize*500: 330.956ms
benchNoirV2CensorFunction*500: 26.536ms
benchFastRedactCensorFunction*500: 42.226ms
benchNoirV2*500: 52.377ms
benchFastRedact*500: 8.433ms
benchFastRedactRestore*500: 10.723ms
benchNoirV2Wild*500: 21.983ms
benchFastRedactWild*500: 35.885ms
benchFastRedactWildRestore*500: 40.448ms
benchFastRedactIntermediateWild*500: 186.271ms
benchFastRedactIntermediateWildRestore*500: 115.239ms
benchJSONStringify*500: 289.707ms
benchNoirV2Serialize*500: 395.613ms
benchFastRedactSerialize*500: 318.428ms
benchNoirV2WildSerialize*500: 356.309ms
benchFastRedactWildSerialize*500: 332.504ms
benchFastRedactIntermediateWildSerialize*500: 410.139ms
benchFastRedactIntermediateWildMatchWildOutcomeSerialize*500: 546.796ms
benchFastRedactStaticMatchWildOutcomeSerialize*500: 362.284ms
benchNoirV2CensorFunction*500: 14.827ms
benchFastRedactCensorFunction*500: 37.329ms
```

The results look within noise to me, which I assume is because this PR is now pretty aggressive about not constructing/passing the path if censor isn't a function and (a new check) if its `.length` isn't greater than 1

For the wholly-new benchmarks, here are the results I got:

```
benchFastRedactCensorFunctionIntermediateWild*500: 157.490ms
benchFastRedactCensorFunctionWithPath*500: 28.350ms
benchFastRedactWildCensorFunctionWithPath*500: 93.050ms
benchFastRedactIntermediateWildCensorFunctionWithPath*500: 128.482ms
...
...
benchFastRedactCensorFunctionIntermediateWild*500: 147.375ms
benchFastRedactCensorFunctionWithPath*500: 12.012ms
benchFastRedactWildCensorFunctionWithPath*500: 88.490ms
benchFastRedactIntermediateWildCensorFunctionWithPath*500: 127.904ms
```

If all those benchmarks look good, I'll update the tests.